% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.anomalies.r
\name{create_anomalies}
\alias{create_anomalies}
\title{Create Anomaly File Function}
\usage{
create_anomalies(var.name, series.file, read.dir, anomaly.file, write.dir,
  monthly.clim, buffer.edge)
}
\arguments{
\item{var.name}{Variable name for the netcdf file.}

\item{series.file}{File name for input netcdf file.}

\item{read.dir}{Location for the input file.}

\item{anomaly.file}{File name for anomaly netcdf file.}

\item{write.dir}{Location for the anomaly file.}

\item{monthly.clim}{Monthly climatologies for the baseline.}

\item{buffer.edge}{Fill missing data at edges DEFAULT=NULL}
}
\description{
Calculates monthly anomalies of daily values relative to
a climatological period.
}
\keyword{anomalies}
