% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.daily.scale.r
\name{daily_target_scale}
\alias{daily_target_scale}
\title{Daily Target Scale Function}
\usage{
daily_target_scale(var.name, interp.file, interp.dir, target.file, target.dir,
  downscale.dir)
}
\arguments{
\item{var.name}{Variable name for the netcdf file.}

\item{interp.file}{Interpolated anomaly file.}

\item{interp.dir}{Location for the interpolated files.}

\item{target.file}{Target climatology file.}

\item{target.dir}{Location for the target climatology file.}

\item{downscale.dir}{Location for the downscaled file.}
}
\description{
Iterates over the yearly anomaly files and bias-corrects them
to the target resolution.
}
\keyword{downscale}
