% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.anomalies.r
\name{edge_buffer}
\alias{edge_buffer}
\title{Extend Area Edge for Interpolation Function}
\usage{
edge_buffer(data, n = 3)
}
\arguments{
\item{data}{Input anomaly data}

\item{n}{Defaults to 3.}
}
\description{
Repeats values at the edge of the file area to add a buffer
for the interpolation step. Only necessary for edges with
missing data.
}
\keyword{buffer}
