% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.anomalies.r
\name{find_interval_dates}
\alias{find_interval_dates}
\title{Find Interval Dates Function}
\usage{
find_interval_dates(dates, yst, yen)
}
\arguments{
\item{dates}{PCICt time series.}

\item{yst}{Start year}

\item{yen}{End year}
}
\description{
Find the start and end dates for a time series given an 
interval with a start and end year.
}
\keyword{interval}
