% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.anomalies.r
\name{interpolate_anomalies}
\alias{interpolate_anomalies}
\title{Interpolate Anomalies Function}
\usage{
interpolate_anomalies(var.name, anomaly.file, read.dir, grid.file, grid.dir,
  write.dir)
}
\arguments{
\item{var.name}{Variable name for the netcdf file.}

\item{anomaly.file}{File name for anomaly netcdf file.}

\item{read.dir}{Location for the anomaly file.}

\item{grid.file}{File containing target grid information.}

\item{grid.dir}{Location for the grid file.}

\item{write.dir}{Location for the interpolated files.}
}
\description{
Uses CDO to bilinearly interpolate calculated anomalies to a
target resolution. The interpolated files are split into 
individual years to reduce their size.
}
\keyword{interpolate}
