% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.anomalies.r
\name{make_anomalies}
\alias{make_anomalies}
\title{Make Anomalies Function}
\usage{
make_anomalies(var.name, series.file, read.dir, anomaly.file, write.dir, yst,
  yen, buffer.edge = NULL)
}
\arguments{
\item{var.name}{Variable name for the netcdf file.}

\item{series.file}{File name for input netcdf file.}

\item{read.dir}{Location for the input file.}

\item{anomaly.file}{File name for anomaly netcdf file.}

\item{write.dir}{Location for the anomaly file.}

\item{yst}{Climatology starting year}

\item{yen}{Climatology ending year}

\item{buffer.edge}{Fill missing data at edges DEFAULT=NULL}
}
\description{
Calculates monthly anomalies of daily values relative to
a climatological period.
}
\keyword{anomalies}
