% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.anomalies.r
\name{pcict_time}
\alias{pcict_time}
\title{File Time Series Function}
\usage{
pcict_time(nc)
}
\arguments{
\item{nc}{ncdf4 object of opened netcdf file.}
}
\description{
Reads the netcdf file time dimensions as a PCICt time
object and returns a PCICt time series
}
\keyword{time}
